enable fastpath in 2.8
