+++ Makefile.in
@@ -1,8 +1,8 @@
-px-objs = pxd.o dev.o iov_iter.o px_version.o pxd_fastpath.o
+px-objs = pxd.o dev.o iov_iter.o px_version.o pxd_fastpath.o kiolib.o pxd_bio_makereq.o
 obj-m = px.o
 
 KBUILD_CPPFLAGS := -D__KERNEL__
-PXDEFINES := -D__PXKERNEL__ -D__PX_FASTPATH__
+PXDEFINES := -D__PXKERNEL__ -D__PX_FASTPATH__ -D__PXD_BIO_MAKEREQ__
 
 KVERSION=$(shell uname -r)
