+++ fuse_i.h
@@ -100,7 +100,7 @@ struct fuse_req {
 
 	/** Request completion callback */
 	/** return: 'true' to free the request, 'false' otherwise */
-	bool (*end)(struct fuse_conn *, struct fuse_req *, int status);
+	bool (*end)(struct fuse_conn *, struct fuse_req *);
 
 	/** Associate request queue */
 	struct request_queue *queue;
